Bump version
Making this another major change 0.1 -> 0.2 as it indicates a change of intention in how the models will be integrated with apps. The expectation is that all events triggered by workflow changes (eg. emailing, indexing, etc) will be handled by the app where the change is triggered. The models remain responsible for the consistency of the data and its internal handling but not how it interfaces with the outside world.