Minor test changes to ease testing of new taggable API
Also marked a few tests as pending because there's a bit more thought needed as to the correct behaviour.

eg. when a primary section is unset but we have three other sections, one of those other sections would become primary. I think that's right but would like to discuss it.