Tidy up ArtefactTagTest
This used to be where all the tests for our tagging behaviour lived. That is all tested elsewhere now but this test lives on to make sure we cover any layers we use in Artefact around the tagging behaviour. For now that's the 'section' method which is needed to maintain compatibility with current api/front-end behaviour.