Add SectionTag class and modify repository so that loading a section gives us a SectionTag
This is a somewhat hacky implementation and mongoid would do it for us if we were using the _type field rather than the tag_type field, but it has the advantage of working without needing a data migration.

We're adding sub-classes for tags so we can begin to attach specific functionality to types of tags. We should use this minimally, but it'll be useful as there are a few business rules to apply to specific types of tags.