SectionTags can have parents
We had been handling this with a rather hacky split statement on the tag_id but shouldn't rely on that.

This change requires a migration which is stored as a rake task (rake migrations:explicit_parent_id). If we're running that anyway we might want to consider renaming the tag_type field to _type which would get rid of the hacky code to set tag class.