Transition between GuideEdition and AnswerEdition
As a first step towards allowing content editors to convert
AnswerEditions to GuideEditions (and vice versa), make sure that
the Edition class has the relevant transitions baked in to start with.

Tests put in place to make sure that the relevant content changes occur
to go from a non-parted Edition to a parted Edition (and vice versa).