Add a new Edition called VideoEdition
This is a new kind of Edition which will hold the Business Link
video content. For the time being these pages are very basic
and will only hold a URL and some summary text.

The reason for creating a new Edition kind specifically rather than
using the (already present) video view for GuideEdition is that
this isn't a parted Edition and doesn't require as much content
to run. Futhermore, these are self-contained pieces of content
rather than belonging to a larger piece as the current uses of
videos on Guides exist to do.