Tags can have short descriptions
As part of the work to enable the homepage (frontend) to populate the
sections it has we need tags to have short descriptions.

This is because the current descriptions, while useful, are too long to
fit on the homepage. As a result these tags will have a shorter
description to go with their long-form description.