Monkey patch mongoid to allow concrete tag classes based on tag_type
We're using tags in a variety of ways and indicating the types using
the tag_type field. We want it to be easy to materialise those tags
as objects of specific classes, not just the Tag class. This introduces
a monkey patch for mongoid to make that easier, bringing mongoid closer
to ActiveRecord in that respect.