Nail down multi_json version to 1.3.7
multi_json 1.4.0 was released on Friday November 30th. It looks like
activesupport, which has a dependency on multi_json (~> 1.0), is pulling
in the new version, but gemfury has a dependency on multi_json (< 1.4).
This causes a conflict which prevents bundler from working at all.

We tried specifying multi_json (< 1.4) in the gemspec, but this doesn't
seem to be enough to convince bundler to pull in an older version and so
the conflict continues.

We therefore specify the exact version 1.3.7, which was the last version
before 1.4.

Worked on this with @jordanhatch.