Correctly validate Inside Gov slugs
Clashing Inside Government slugs can be differentiated by automatically
appending '--$version' to their values. Prior to this commit this type
of slug, which is valid and in public use, wouldn't pass validation.

We now special case that type of slug. As a bonus there are tests for
the edge cases in the SlugValidator class.