Don't change the slug during validation
As the slug for an artefact was being passed as a reference into
the slug validator, slugs for detailed guides were being changed
during validation when we replaced any double dashes with a single
dash.

This makes sure that, when we pass the slug value directly into the
validator, we clone it so that any changes don't affect the value
in the Artefact.