Remove requires conflicting with Rails Engine load
Rails engine eager_load! loads the slug_validator
anyways, so we should not try to load it again.

When the model artefact.rb requires slug_validator
and the Rails engine load doesn't yet know of it,
it tries loading it again. This was causing the
InstanceValidator to be redefined with a new
instance of Struct causing superclass mismatch
error.