Add RenderedSpecialistDocument model
This will hold a rendered version of the specialist document, including
the body in HTML format.

The rendered version will be created/updated by the specialist publisher
every time the document is published.

This is so that we don't need to add support for InlineAttachment tags
into the general-purpose govspeak helper, nor do we need to give
govuk_content_api knowledge about attachments. Instead the rendering of
the HTML will be self-contained within the specialist-publisher
application.

This fits in with the new architectural patterns envisaged Brad/Alex
Tomlins (I've discussed with Alex).