Remove parent_id validation in tag_ids
Validation the tag_id for the presence/lack of the parent_id will 
require many changes to tests across multiple apps, as many tests 
create stub data with a child-like tag id, but without any parent_id.

We should still do this at some point, but for now a simpler validation 
for a single slash will go enough of the way to ensuring the quality of
tag IDs.