Remove fact check from the Edition model.
This isn't a concern of the shared model: only Publisher should need to
care about the logic behind fact check addresses, especially since it's
becoming configuration-dependent.

This does change one piece of functionality: the default message when
something goes out for fact check doesn't include the response address.
This isn't a problem (as discussed with @lutgendorff), because this
address is exposed elsewhere in the same Publisher view.