On save, only update slug for edition from artefact
As the section, department and business_proposition fields are 
deprecated, there’s not a good reason to denormalize these to an edition
whenever an artefact is saved.

The canonical source of the name of an edition is Edition#title, not 
Artefact#name [1] so it doesn’t make sense to denormalize this back 
again either.

[1] https://github.com/alphagov/govuk_content_api/blob/bf9ea854/lib/presenters/minimal_artefact_presenter.rb#L24