Only update edition slug when draft artefacts are saved
The only time when a slug can be changed for an artefact is when its in
draft, therefore we only need to update the edition for a draft artefact.

In addition, attemping to edit many fields on a published edition raises
an error as they cannot be edited. This will prevent these from
happening.