Denormalising users shouldn't call save on edition
www.agileplannerapp.com/boards/173808/cards/4296

Denormalising users is a process that's setting
values for a few columns, which doesn't need the
the entire validation chain or callbacks to
execute. there are no validations related to
denormalisation, all the more reason for it
to be a plain #set.

there's a cost involved here of multiple
update queries being fired, but that should
be substantially lower than triggering callbacks
and validation chain repeatedly.