Delete section field on editions
www.agileplannerapp.com/boards/173808/cards/5444

there's this bug reported in Publisher where
someone's adding a section to an artefact in
Panopticon, yet the section is not showing up
in the documents table in Publisher. this is
because the editor adds edition.artefact.section
whereas Publisher shows the edition.section field.
these 2 fields don't get synced after an edition
is created from artefact metadate.

here's a proposal to delete the editon.section
field altogether, so that we reduce coupling
between these applications and not have to
worry about duplicating these fields across
editions and their artefacts.

it seems that the Publisher documents table is
the only place we use edition.section. all the
more reason to switch that to use section from
the artefact, and drop the edition.section.

taking an example of https://www.gov.uk/payslips.
this edition has a section: "Work:Employment rights"
its artefact has section: "Working, jobs and pensions:
Your pay, tax and the National Minimum Wage". on
frontend, the artefact's section is used to create
breadcrumbs and report Section variable to Google
Analytics. only Publisher shows the edition.section.