Validate only govspeak fields for safe html
www.agileplannerapp.com/boards/173808/cards/5154

looking at the HTML validator in govspeak suggests
that it is useful to validate that HTML generated
from govspeak is safe. however, these checks seem
unnecessary for non-govspeak fields, because rails
ensures that safe html is rendered, and we take
care to not mark user input as `.html_safe`.

this is blocking editors from adding html in the
edition notes, which should be allowed.