Add methods to Edition to return major updates
This changes the Edition model to include methods to return 
major updates in the edition series. It also includes methods to
return the latest change note for the series, and the 'updated_at' 
timestamp for the latest major update.

Some of this behaviour previously existed in the 'enhancements' of the 
Edition model in Publisher.