Define GOVSPEAK_FIELDS only on models that have it
the safe_html_validator_test mandates all models to
validate using the SafeHtml validator irrespective
of whether they had any govspeak or not. this in
turn requires all models to define GOVSPEAK_FIELDS.

i've updated the test to only validate that models
that define this constant validate using SafeHtml.
models have better reason to define GOVSPEAK_FIELDS
in that it helps content api identify which fields
need to be rendered as govspeak, so it is not likely
to get missed.