Use ID and order to key part validation errors
Relying on the order parts arrived in led to ambiguity in determining
which error belonged to which part when parts were being re-arranged;
re-arranged parts were not reflected in the error response.

Instead be explicit about the part that’s failing validation by keying
on both ID and order. Use order because on the client side new parts
will not have a known ID and order will be used to identify the failing
part.