Allow optional content on completed transaction pg
https://trello.com/c/caDndyIX

As a Service Manager
I want to be able to indicate whether Organ Donor Registration promotion
  should be incorporated into my "Completed transaction" page
so that I can offer users of my service to register for Organ Donation

In order to avoid adding more fields to Edition for such requirements,
we're storing information on one Edition field: presentation_toggles.
It is a Hash containing a boolean which toggles display of this field,
and a URL which the promotion links to. Going forward if we have more
such requirements, we can dump it into the presentation_toggles Hash.

PresentationToggles includes that new field and adds getter/setter methods
on the Edition, to make it behave as if individual PresentationToggles are
multiple fields on the Edition. This helps with form submissions.