Fix downtime test which fails with latest versions of gems
When I checked the gem out on my local machine and ran the tests I saw the
following failure:

      1) Failure:
    test: publicising downtime should start at midnight a day before it is scheduled. (DowntimeTest) [test/models/downtime_test.rb:72]:
    Failed assertion, no message given.

I assume this was failing when the CI build is not, because the CI build has
not been run for 25 days and new versions of some of the less-constrained
dependent gems are now available and installed by bundler.

I struggled to work out why the test was failing, because the test was using
instances of `Date` and `Time` whereas `Downtime#publicise?` was operating on
instances of `TimeWithZone`, `DateTime` & `Date`.

So I updated the test to only use instances of `TimeWithZone` and to be more
deterministic and the test started passing.