Relax constraint on govspeak dependency.
Given that GDS is in control of govspeak and it's supposed to use semantic
versioning, it ought to be safe to relax this constraint from `~> 3.1.0` to
`~> 3.1`.

I'd like this change because I'm running into a problem running the tests in
panopticon due to an incompatibility between nokogiri and libxml on Mac OS X.
Unfortunately this constraint is preventing me from upgrading nokogiri within
panopticon.

I see that this change was attempted [1] some time ago, but then immediately
reverted [2], so perhaps there are some other problems of which I'm unaware.

[1]: https://github.com/alphagov/govuk_content_models/commit/6403139e1e0f627d0c55957d75e446ef3389976e
[2]: https://github.com/alphagov/govuk_content_models/commit/32b9cbcf3da0089a0edf0c247bf437ea1374744b