Simplify changing Edition types
The old code had lots of hard-coded bespoke conditions that examined the types
of the objects to decide what to copy. This changes it to use duck typing to
figure out what data can be copied between types.

This approach is shorter, simpler and handles more conversions and will handle
new types without them having to be explicitly added. It also reveals the
intent better.