Include SimpleSmartAnswer nodes in whole body
This change makes SimpleSmartAnswer nodes consistent with Parted documents
(Guides and Programmes).

whole_body is used in two contexts:
* when displaying the diff between two editions in Publisher
* when populating a new edition when changing the format

Because we weren't including the nodes in the whole_body, they weren't showing
up in Publisher's diff, even though most changes occur in the nodes.

If you changed to a different edition type, the nodes wouldn't be copied.