Extract whole_body-setting logic to method
This new method lives on the target edition of the cloning process, inverting
the resposibility of the `respond_to?` and value-setting.

I think a more object-oriented approach would be to use polymorphism and let
the target edition's own class (i.e. not its superclass, Edition) decide how to
handle this task. I hope this is a decent step in that direction.

I worry that moving the handling of this task to those subclasses of Edition
means spreading this cloning mess even wider into this application, while
ideally cloning an Edition would be considered a distinct responsibility
contained in its own object.