When archiving an artefact, allow invalid editions to be archived
Before this change, attempting to archive an artefact with invalid editions
would successfully mark the artefact as archived, but the editions would be
left in their state.

We have published editions which are invalid because validation rules have been
introduced since they were last worked on/saved.

Ideally, the state machine transition for archiving would skip validations, but
state_machine doesn't seem to support this.