Add "official_statistics" format as owned by Whitehall
Recently Statistics were relabelled in Whitehall as Official
Statistics, including the parameterised document format name sent to
Panopticon. As Panopticon didn’t know about the format, the validations
were failing when Whitehall attempted to send the artefacts, causing
errors for Whitehall sidekiq jobs.

This PR adds the `official_statistics` format to `Artefact` for
Panopticon to pick up. Further investigation is required to determine
whether the old `statistics` format should be removed yet, or if a data
migration should be performed in Panopticon first to change the format
of the artefacts.