Remove _destroy parameter when updating.
The overridden `update_attributes` method in SimpleSmartAnswer passes all
the attributes onto the individual Node.update_attributes calls, including the
_destroy parameter which causes an error because it is not a defined field.
Remove it first.

Note that the problem linked in the comment (updated to point at the new Jira
bug tracker) is claimed to be fixed, but still occurs if the overridden method
is removed.