Save artefact automatic actions with save_as_task method
Given that during the incident in 2016-02-14 it was not clear what meant
"An unknown user".
When an artefact action has no user it means it was updated by one of
the tasks, in panopticon it would be one of these three:

TaggingUpdater
OrganisationSlugChanger
rake task "Move content to new topics"
In order to know which one, we have agreed to add a new field in
ArtefactAction called "task_performed_by" that receives a string with
the name of the automatic task.

This will facilitate future debugging and more visibility to developers
if future incidents happen.