Add validations for promotion choice
Promotion choice has to be from the list (none, organ_donor,
register_to_vote) and the url must be present if the choice is not none.
We've left the old validations for the legacy organ donor storage scheme
for now as while the promotion_choice API falls back to the old one, the
old scheme doesn't "fall forward" to the new one.  A later PR will remove
the old scheme entirely after a data-migration.