Implement "fall back" for legacy promotions correctly
When checking for the "promotion_choice" key inside the
presentation_toggles hash field if it wasn't present we would return the
default hash.  However this wouldn't be stored in the model so any changes
that were written to it would not be persisted.  To fix we write the
default hash into the presentation_toggles if the key wasn't already there.