Handle legacy scheme for promos more robustly
In 46f2e931cf8373efa59a9e187f0c7b18e7b1f6d8 we changed how we handled
missing "promotion_choice" keys in the presentation_toggles field to
better deal with existing documents that didn't have that key because
they were created before the new defaults were in place.  Our solution
used ``||=`` to store the default values in the presentation_toggles
hash for the key if it wasn't already present.

Our test for this worked with an unsaved object which meant that the
presentation_toggles field was still an instance of Hash.  Once it's
saved to the DB however and retrieved it becomes a BSON::Document and
this has slightly different semantics for ``||=`` assignment.

With a hash:

    a = {}
    b = {foo: {bar: "baz"}}
    c = (a[:foo] ||= b[:foo])
    c.object_id == a[:foo].object_id

Then the final line returns `true`.  However if `a` in this example
is a BSON::Document and not a Hash then the final line returns `false`.
This means that anything we do to `c` isn't actually acting on the
data stored in `a[:foo]`, but on the value we get from `b[:foo]`.

Having worked this out we change the test to work on a saved instance
to trigger the failure.  Note that to properly construct an instance
without `promotion_choice` in presentation_toggles we need to save
without validations as one of our validations would end up populating
the `promotion_choice` key and this would mean our test wouldn't be
a true example of real world data.

The solution is then to unroll the ``||=`` and do the assignment and
return in two steps, ensuring that we only expose the data stored in
presentation_toggles rather than the data we wanted to put in.