Define start button text for SimpleSmartAnswerEdition
As requested in this trello card https://trello.com/c/CUe4odRB/45-make-start-now-editable-in-simple-smart-answers
the need to add the capability to change the text of the start button for simple smart answers.

Here, I have defined a new attribute "start_button_text" with type of string and a default text of "Start Now"

Also included are tests that ensure that when the button text is change the change is recorded and part of a valid
SimpleSmartAnswerEdition object.

Add validation for start_button_text