Remove the LocalInteraction model
We stop importing these from local directgov admin because we've built
[Local Links Manager](https://github.com/alphagov/local-links-manager)
to act as the source of truth for these.  This means we no longer need
this model.  We leave LocalAuthority and LocalService as, although LLM is
the source of truth for these as well, these models are still used in
places (Publisher needs LocalService for validating local_transaction
editions, and we use LocalAuthority for licenses).