Add require_dependency stub for tests
The test environment for this gem does not load a full rails environment
and so require_dependency is not available.  Adding a require statement to
load in the part of rails that provides it (active_support/dependencies)
causes the exact errors we want require_dependency to prevent, so it's
possible that this is also not enough of rails to make things work
correctly.  To avoid this and allow the tests to pass we define our own
require_dependency method for the test environment that is a pass-through
to a normal require.