Remove whitehall references
Any code related to Whitehall is not needed anymore. Originally this code was added to support Whitehall content being registered in Panopticon for tagging purposes. Panopticon has been retired and tagging now happens in content tagger, which does not use this gem so we're safe to remove the code.

Original PR introducing Whitehall formats
https://github.com/alphagov/govuk_content_models/pull/58

Commit that creates the current hash structure
https://github.com/alphagov/govuk_content_models/commit/c3056e1bd14bffc03be292b4e6e59e034acdc6e8